Automation/requirements/requirements.txt
adal==1.2.5
altgraph==0.17
antlr4-python3-runtime==4.7.2
appdirs==1.4.4
applicationinsights==0.11.9
argcomplete==1.12.1
astroid==2.2.5
azure-appconfiguration==1.1.1
azure-batch==9.0.0
azure-common==1.1.25
azure-core==1.8.2
azure-cosmos==3.2.0
azure-datalake-store==0.0.51
azure-devops==6.0.0b1
azure-functions-devops-build==0.0.22
azure-graphrbac==0.60.0
azure-identity==1.5.0
azure-keyvault-administration==4.0.0b2
azure-keyvault-secrets==4.2.0
azure-loganalytics==0.1.0
azure-mgmt-advisor==2.0.1
azure-mgmt-apimanagement==0.2.0
azure-mgmt-appconfiguration==0.6.0
azure-mgmt-applicationinsights==0.1.1
azure-mgmt-authorization==0.61.0
azure-mgmt-batch==9.0.0
azure-mgmt-batchai==2.0.0
azure-mgmt-billing==0.2.0
azure-mgmt-botservice==0.2.0
azure-mgmt-cdn==5.0.0
azure-mgmt-cognitiveservices==6.3.0
azure-mgmt-compute==14.0.0
azure-mgmt-consumption==2.0.0
azure-mgmt-containerinstance==1.5.0
azure-mgmt-containerregistry==3.0.0rc15
azure-mgmt-containerservice==9.4.0
azure-mgmt-core==1.2.0
azure-mgmt-cosmosdb==1.0.0
azure-mgmt-datalake-analytics==0.2.1
azure-mgmt-datalake-nspkg==3.0.1
azure-mgmt-datalake-store==0.5.0
azure-mgmt-datamigration==0.1.0
azure-mgmt-deploymentmanager==0.2.0
azure-mgmt-devtestlabs==4.0.0
azure-mgmt-dns==2.1.0
azure-mgmt-eventgrid==3.0.0rc7
azure-mgmt-eventhub==4.1.0
azure-mgmt-hdinsight==1.7.0
azure-mgmt-imagebuilder==0.4.0
azure-mgmt-iotcentral==3.0.0
azure-mgmt-iothub==0.12.0
azure-mgmt-iothubprovisioningservices==0.2.0
azure-mgmt-keyvault==7.0.0b3
azure-mgmt-kusto==0.3.0
azure-mgmt-loganalytics==0.7.0
azure-mgmt-managedservices==1.0.0
azure-mgmt-managementgroups==0.2.0
azure-mgmt-maps==0.1.0
azure-mgmt-marketplaceordering==0.2.1
azure-mgmt-media==2.2.0
azure-mgmt-monitor==0.11.0
azure-mgmt-msi==0.2.0
azure-mgmt-netapp==0.12.0
azure-mgmt-network==12.0.0
azure-mgmt-nspkg==3.0.2
azure-mgmt-policyinsights==0.5.0
azure-mgmt-privatedns==0.1.0
azure-mgmt-rdbms==3.1.0rc1
azure-mgmt-recoveryservices==0.4.0
azure-mgmt-recoveryservicesbackup==0.6.0
azure-mgmt-redhatopenshift==0.1.0
azure-mgmt-redis==7.0.0rc2
azure-mgmt-relay==0.1.0
azure-mgmt-reservations==0.6.0
azure-mgmt-resource==10.2.0
azure-mgmt-search==2.1.0
azure-mgmt-security==0.4.1
azure-mgmt-servicebus==0.6.0
azure-mgmt-servicefabric==0.5.0
azure-mgmt-signalr==0.4.0
azure-mgmt-sql==0.21.0
azure-mgmt-sqlvirtualmachine==0.5.0
azure-mgmt-storage==11.2.0
azure-mgmt-synapse==0.3.0
azure-mgmt-trafficmanager==0.51.0
azure-mgmt-web==0.48.0
azure-multiapi-storage==0.5.2
azure-nspkg==3.0.2
azure-storage-common==1.4.2
azure-synapse-accesscontrol==0.2.0
azure-synapse-artifacts==0.3.0
azure-synapse-spark==0.2.0
bcrypt==3.2.0
beautifulsoup4==4.9.0
cachetools==4.2.0
certifi==2020.12.5
cffi==1.13.2
cfgv==3.2.0
chardet==3.0.4
click==7.1.2
colorama==0.4.1
cryptography==2.8
distlib==0.3.1
example==0.1.0
fabric==2.5.0
filelock==3.0.12
future==0.18.2
google-api-core==1.24.1
google-auth==1.24.0
google-cloud-core==1.5.0
google-trans-new==1.1.9
googleapis-common-protos==1.52.0
grpcio==1.34.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2020.12.22
httpcore==0.9.1
httpx==0.13.3
humanfriendly==8.2
hyperframe==5.2.0
identify==1.4.25
idna==2.8
importlib-metadata==1.7.0
invoke==1.4.1
isodate==0.6.0
isort==4.3.20
javaproperties==0.5.1
Jinja2==2.11.2
jmespath==0.10.0
jsmin==2.2.2
jsondiff==1.2.0
knack==0.7.2
lazy-object-proxy==1.4.1
libcst==0.3.16
lxml==4.3.4
MarkupSafe==1.1.1
mccabe==0.6.1
mock==4.0.2
msal==1.7.0
msal-extensions==0.3.0
msedge-selenium-tools==3.141.2
msrest==0.6.10
msrestazure==0.6.4
mypy-extensions==0.4.3
nodeenv==1.4.0
oauthlib==3.1.0
packaging==20.4
paramiko==2.7.2
pefile==2019.4.18
Pillow==6.1.0
pkginfo==1.6.1
pluggy==0.13.1
portalocker==1.7.1
pre-commit==2.6.0
proto-plus==1.13.0
protobuf==3.14.0
psutil==5.7.3
py==1.8.1
py-googletrans==1.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.19
Pygments==2.7.2
PyInstaller==3.6
PyJWT==1.7.1
pylint==2.3.1
PyNaCl==1.4.0
pynput==1.4.2
pyOpenSSL==19.1.0
pyparsing==2.4.7
pypiwin32==223
pyreadline==2.1
PyRECONSTRUCT==2.2.0
python-dateutil==2.8.1
python-docx==0.8.10
pyttk==0.3.2
pytz==2019.1
pywin32==224
pywin32-ctypes==0.2.0
PyYAML==5.3.1
requests==2.22.0
requests-oauthlib==1.3.0
rfc3986==1.4.0
rsa==4.6
scp==0.13.3
selenium==3.141.0
six==1.15.0
sniffio==1.2.0
soupsieve==2.0
sshtunnel==0.1.5
tabulate==0.8.7
toml==0.10.1
tox==3.19.0
typed-ast==1.3.5
typing-extensions==3.7.4.3
typing-inspect==0.6.0
urllib3==1.25.3
virtualenv==20.0.30
virtualenvwrapper-win==1.2.5
vsts==0.1.25
vsts-cd-manager==1.0.2
websocket-client==0.56.0
wrapt==1.11.1
xmltodict==0.12.0
zipp==3.1.0
